+++ dymension/libs/kaspa/lib/api/src/models/blue_score_response.rs
@@ -14,7 +14,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct BlueScoreResponse {
     #[serde(rename = "blueScore", skip_serializing_if = "Option::is_none")]
-    pub blue_score: Option<i32>,
+    pub blue_score: Option<i64>,
 }
 
 impl BlueScoreResponse {
